+++ R/colorPalette.R
@@ -4,8 +4,8 @@
 #'
 #' @usage colorPalette(name, n = NULL)
 #'
-#' @param name   name of the color palette. Can be \code{random} for random colors or \code{complement} for complementing colors, but can also be the name of a pre-implemented palette. See the \code{details} section for a list of pre-implemented palettes.
-#' @param n      the number of colors to select from the palette. Required if \code{name = 'random'} or \code{name = 'complement'}. Otherwise, if \code{NULL}, automatically selects all colors from the chosen palette.
+#' @param name   name of the color palette. Can be \code{random} for random colors, \code{complement} for complementing colors, \code{divergent} for equally spaced colors, or \code{random-palette} for a random palette, but can also be the name of a pre-implemented palette. See the \code{details} section for a list of pre-implemented palettes.
+#' @param n      the number of colors to select from the palette. Required if \code{name = 'random'}, \code{name = 'complement'}, or \code{name = 'divergent'}. Otherwise, if \code{NULL}, automatically selects all colors from the chosen palette.
 #'
 #' @details The following color palettes are implemented:
 #'
@@ -17,7 +17,7 @@
 #' @author Koen Derks, \email{koen-derks@hotmail.com}
 #'
 #' @examples
-#' colorPalette("complement", 5)
+#' colorPalette("divergent", 5)
 #' @keywords canvas palette
 #'
 #' @export
@@ -53,26 +53,50 @@ colorPalette <- function(name, n = NULL) {
         palette[i] <- .hsl_to_rgb(h = tmp, stats::runif(1, .4, .8), stats::runif(1, .4, .8))
       }
     }
+  } else if (name == "divergent") {
+    palette <- character(n)
+    start <- stats::runif(1, 0, 260)
+    colSeq <- seq(from = 1, to = 360, length.out = n)
+    h <- (colSeq + start) %% 360
+    for (i in 1:length(palette)) {
+      palette[i] <- .hsl_to_rgb(h = h[i], stats::runif(1, .4, .7), stats::runif(1, .4, .7))
+    }
   } else {
+    if (name == "random-palette") {
+      name <- sample(c(
+        "blackwhite", "dark1", "dark2", "dark3", "flora", "gogh", "house", "jasp", "jfa", "jungle",
+        "klimt", "kpd", "lava", "nature", "neon1", "neon2", "origami", "retro1", "retro2", "retro3", "sooph",
+        "tuscany1", "tuscany2", "tuscany3", "vrolik1", "vrolik2", "vrolik3"
+      ), size = 1)
+    }
     palette <- switch(name,
       "blackwhite" = c("black", "white"),
       "dark1" = c("#161616", "#346751", "#C84B31", "#ECDBBA"),
       "dark2" = c("#1B262C", "#0F4C75", "#3282B8", "#BBE1FA"),
       "dark3" = c("#222831", "#393E46", "#00ADB5", "#EEEEEE"),
+      "flora" = c("#000000", "#f2f2eb", "#ccb77c", "#523402"),
+      "gogh" = c("#8699b5", "#161918", "#9d8018", "#232d8a", "#b4ad5a", "#c2ccd5"),
       "house" = c("#191919", "white", "#960606", "#cca222", "#036440"),
       "jasp" = c("#14a1e3", "#ffffff", "#f7971c", "#8cc63e", "#000000"),
       "jfa" = c("#2a94d1", "#e89643", "#233e4a"),
       "jungle" = c("#4e6349", "#614128", "#171812", "#698144", "#917861"),
+      "klimt" = c("#847049", "#ba9d3e", "#29241f", "#a787b0", "#cd5627", "#7ea36e"),
+      "kpd" = c("#c42f32", "#315b99", "#354741"),
       "lava" = c("#f59907", "#a42300", "#482a22", "#050000", "#6b0800"),
       "nature" = c("forestgreen", "dodgerblue", "brown", "white", "gray"),
       "neon1" = c("#F7FD04", "#F9B208", "#F98404", "#FC5404"),
       "neon2" = c("#F5F7B2", "#1CC5DC", "#890596", "#CF0000"),
+      "origami" = c("#01364f", "#84231e", "#247c86", "#e8674d", "#dfdbbe", "#fdf4b4"),
       "retro1" = c("#0A1931", "#185ADB", "#FFC947", "#EFEFEF"),
       "retro2" = c("#DDDDDD", "#222831", "#30475E", "#F05454"),
       "retro3" = c("#111D5E", "#C70039", "#F37121", "#C0E218"),
+      "sooph" = c("#d6c398", "#cbabdb", "#139485", "#9e1710", "#414952"),
       "tuscany1" = c("firebrick", "goldenrod", "forestgreen", "navyblue"),
       "tuscany2" = c("#500342", "#023b59", "#f9efdd", "#deaa70", "#711308"),
-      "tuscany3" = c("#b08653", "#f5daba", "#c9673c", "#f2ab4e", "#a1863b")
+      "tuscany3" = c("#b08653", "#f5daba", "#c9673c", "#f2ab4e", "#a1863b"),
+      "vrolik1" = c("#4ca787", "#183867", "#ea8857", "#442a37", "#ffb747"),
+      "vrolik2" = c("#d8c888", "#f4edf3", "#fa428c", "#1900b4", "#ecd07d"),
+      "vrolik3" = c("#774f38", "#e08e79", "#f1d4af", "#ece5ce", "#c5e0dc")
     )
     if (is.null(palette)) {
       stop(paste0("'", name, "' is not an existing palette"))
@@ -81,7 +105,8 @@ colorPalette <- function(name, n = NULL) {
       n <- length(palette)
     }
     if (n > length(palette)) {
-      stop("attempt to select more colors than are available in this palette")
+      warning("attempt to select more colors than are available in this palette, returning the requested palette with the maximum number of colors")
+      return(palette)
     }
     palette <- palette[1:n]
   }